+++ src/test/fixtures/launchTaskFixture.ts
@@ -1,6 +1,6 @@
 import { TaskDefinition } from 'vscode';
 
-const expectedResult: TaskDefinition = {
+export const debugFixture: TaskDefinition = {
   showDevDebugOutput: 'parsed',
   cwd: '${workspaceRoot}',
   executable: `./build/Clean_project_h7.elf`,
@@ -12,12 +12,26 @@ const expectedResult: TaskDefinition = {
   device: 'stm32h743',
   configFiles: [
     'openocd.cfg',
-
   ],
 };
-export default expectedResult;
 
-export const expectedResultWithSVD = {
-  ...expectedResult,
-  configFiles: [...expectedResult.configFiles, 'STM32H743x.svd']
+export const attachFixture: TaskDefinition = {
+  ...debugFixture,
+  name: 'Attach STM32',
+  request: 'attach',
 };
+
+
+
+export default ([debugFixture, attachFixture]);
+
+export const debugFixtureWithSVD = {
+  ...debugFixture,
+  svdFile: 'STM32H743x.svd',
+};
+
+export const attachFixtureWithSVD = {
+  ...attachFixture,
+  svdFile: 'STM32H743x.svd',
+};
+