+++ src/test/unit/workspaceConfiguration/LaunchConfig.test.ts
@@ -1,13 +1,13 @@
+import { suite, test } from 'mocha';
 
+import {debugFixture} from '../../fixtures/launchTaskFixture';
 import { expect } from 'chai';
-import { test, suite } from 'mocha';
 import getLaunchTask from '../../../configuration/LaunchTasksConfig';
 import { testMakefileInfo } from '../../fixtures/testSTMCubeMakefile';
-import LaunchTestFile from '../../fixtures/launchTaskFixture';
 
 suite('Launch configuration test', () => {
   test('Returns correct config', () => {
     const res = getLaunchTask(testMakefileInfo);
-    expect(res).to.deep.equal(LaunchTestFile);
+    expect(res).to.deep.equal(debugFixture);
   });
 });